Fix applyNumberFormat default value for XLSX
When extract style attribute, a missing "applyNumberFormat" attribute would lead to a bool value being filled with `null`. This value would then be used for comparison (`if ($applyNumberFormat === false)`). This does not take into account the nullable aspect of the value.
There is no reason to keep this value as nullable. If the attribute does not exist, then we should default to "false" (as in "Do not try to apply number format here").